Room safety
`AvailableRoom` was renamed to `Room`, since rooms can now be added to a list of `RoomsInProgress` in the `ChatService`. `Room`s are moved to this array when that are 'started'. Validation checks were added so that players that aren't part of a room can't send a message to that room.