Ending or Disconnecting from a game
Added a method for gracefully ending a game, `EndGame`, but also added a precaution if a player disconnects while in the middle of a game, `OnDisconnectedAsync` and `DisconnectRoom` to find whatever room the disconnected player was in and remove it. It then calls a front-end method (that is yet to be written) that alerts the rest of the players that there was a disconnect.